Organisasjonstype fra @types til api-modul
