+++ src/@types/index.ts
@@ -7,7 +7,6 @@ export type {
 export type { default as Fagsak } from "./fagsak";
 export type { DokumentOversikt, Mottaksretning, FysiskDokument, Dokument, BrevPdfData, SedPdfData } from "./dokument";
 export type { Periode, Lovvalgsperiode, Medlemskapsperiode, OppdaterMedlemskapsperiode } from "./periode";
-export type { default as Organisasjon } from "./organisasjon";
 export type { MedlPeriode } from "./medlemskap";
 export type { default as Aktoer } from "./aktoer";
 export type { Avgiftsperiode, Avgiftsgrunnlag, Avgiftsberegning, AvgiftsgrunnlagInfo } from "./trygdeavgift";
