--- src/@types/organisasjon.ts
@@ -1,10 +1,0 @@
-import { GeneriskAdresse } from "./adresse";
-
-type Organisasjon = {
-  orgnr: string;
-  navn: string;
-  forretningsadresse: GeneriskAdresse;
-  postadresse: GeneriskAdresse;
-};
-
-export default Organisasjon;
