+++ src/felleskomponenter/menypanel/menypunkter/arbeidsgiver_og_virksomhet/arbeidsgiver_og_virksomhet.tsx
@@ -1,7 +1,6 @@
 import React, { ReactNode } from "react";
 import { connect, ConnectedProps } from "react-redux";
 import { RootState } from "AppTypes";
-import { Organisasjon } from "Domene";
 import { AnyAction } from "redux";
 import { ThunkDispatch } from "redux-thunk";
 import { formValueSelector } from "redux-form";
@@ -12,6 +11,7 @@ import * as KV from "../../../../kodeverk";
 import * as Etiketter from "../../etiketter";
 import * as Ikoner from "../../../../resources/images";
 import * as Utils from "../../../../utils";
+import * as Api from "../../../../services/api";
 
 import ArbeidsforholdNorgeListe from "./arbeidsforholdNorgeListe";
 import EditerbartElementListe from "../editerbartElementListe";
@@ -70,7 +70,7 @@ export const ArbeidsgiverOgVirksomhet = ({
   behandlingsgrunnlagEtikett,
 }: ArbeidsgiverOgVirksomhetProps) => {
   const finnOrganisasjon = (orgnr: string) => {
-    const org: Organisasjon = organisasjoner.find((o: Organisasjon) => o.orgnr === orgnr);
+    const org: Api.Types.Organisasjon = organisasjoner.find((o: Api.Types.Organisasjon) => o.orgnr === orgnr);
     return org || { orgnr };
   };
 
@@ -103,7 +103,7 @@ export const ArbeidsgiverOgVirksomhet = ({
         feltNavn="juridiskArbeidsgiverNorge.ekstraArbeidsgivere"
         redigerbart={redigerbart}
         hentOrganisasjon={hentOrganisasjon}
-        transformerOrgTilElement={(org: Organisasjon) => org.orgnr}
+        transformerOrgTilElement={(org: Api.Types.Organisasjon) => org.orgnr}
         findOrganisasjon={finnOrganisasjon}
         defaultElement={null}
         elementerInneholderOrg={(orgListe: string[], orgnr: string) => orgListe.includes(orgnr)}
@@ -124,12 +124,12 @@ export const ArbeidsgiverOgVirksomhet = ({
         feltNavn="selvstendigForetak"
         redigerbart={redigerbart}
         hentOrganisasjon={hentOrganisasjon}
-        transformerOrgTilElement={(org: Organisasjon) => ({ orgnr: org.orgnr })}
-        findOrganisasjon={(org: Organisasjon) =>
-          organisasjoner.find((enkeltOrg: Organisasjon) => enkeltOrg.orgnr === org.orgnr)
+        transformerOrgTilElement={(org: Api.Types.Organisasjon) => ({ orgnr: org.orgnr })}
+        findOrganisasjon={(org: Api.Types.Organisasjon) =>
+          organisasjoner.find((enkeltOrg: Api.Types.Organisasjon) => enkeltOrg.orgnr === org.orgnr)
         }
         defaultElement={{}}
-        elementerInneholderOrg={(orgListe: Organisasjon[], orgnr: string) =>
+        elementerInneholderOrg={(orgListe: Api.Types.Organisasjon[], orgnr: string) =>
           orgListe.find((org) => org.orgnr === orgnr)
         }
       />
