+++ src/felleskomponenter/menypanel/menypunkter/fullmektig/fullmektigRedigeringUtfort.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { Organisasjon } from "Domene";
 
 import * as Nav from "../../../../utils/navFrontend";
 import * as Ikoner from "../../../../resources/images";
@@ -29,7 +28,7 @@ const hentRepresentererTekst = (representererKode: string) => {
 };
 
 interface FullmektigRedigeringUtfortProps {
-  org: Partial<Organisasjon>;
+  org: Partial<Api.Types.Organisasjon>;
   representererKode: string | null;
   kontaktopplysninger: KontaktOpplysning;
 }
@@ -39,8 +38,8 @@ const FullmektigRedigeringUtfort = ({
   representererKode,
   kontaktopplysninger,
 }: FullmektigRedigeringUtfortProps) => {
-  const [kontaktopplysningerOrg] = useAsyncCallbackState(
-    () => Api.Organisasjoner.hentOrganisasjon(kontaktopplysninger.kontaktorgnr),
+  const [kontaktopplysningerOrg] = useAsyncCallbackState<Partial<Api.Types.Organisasjon>>(
+    () => Api.Organisasjoner.hentOrganisasjon(kontaktopplysninger.kontaktorgnr || ""),
     {},
     Utils.logger.error,
     [kontaktopplysninger.kontaktorgnr]
