+++ src/felleskomponenter/stegvelger/stegKomponenter/ftrl/vurderingRepresentant.tsx
@@ -4,7 +4,6 @@ import { change, getFormValues, reduxForm } from "redux-form";
 import { ThunkDispatch } from "redux-thunk";
 import { RootState } from "AppTypes";
 import { Action } from "redux";
-import { Organisasjon } from "Domene";
 
 import * as Nav from "../../../../utils/navFrontend";
 import * as Skjema from "../../../../felleskomponenter/skjema";
@@ -69,7 +68,7 @@ const VurderingRepresentant = ({
 }: Props & PropsFromRedux) => {
   const [representantListe, setRepresentantListe] = useState<Api.Representant.RepresentantListeResDto>([]);
   const [representantData, setRepresentantData] = useState<Api.Representant.RepresentantDataResDto>();
-  const [organisasjon, setOrganisasjon] = useState<Organisasjon | undefined>();
+  const [organisasjon, setOrganisasjon] = useState<Api.Types.Organisasjon | undefined>();
   const hjelpetekstNummer =
     "Representantnummeret du legger til her vil bli overført til Avgiftssystemet (ME7-bildet) når du fatter vedtak.\nSkal du opprette en ny representant, må du gjøre det i Avgiftssystemet.\nListen du finner her oppdateres hvert døgn. Hvis du har opprettet eller endret en representant i Avgiftssystemet i dag, vil du derfor ikke finne oppdateringen her. Dette har ikke betydning for overføringen til Avgiftssystemet, så lenge nummeret er riktig.";
   const hjelpetekstAdresse =
