--- src/services/modules/organisasjoner.js
@@ -1,5 +1,0 @@
-import { cachedGetAsJson } from "../utils";
-import { API_BASE_URL, ORGANISASJONER } from "../api-constants";
-
-// eslint-disable-next-line import/prefer-default-export
-export const hentOrganisasjon = (orgnr) => cachedGetAsJson(`${API_BASE_URL}${ORGANISASJONER}/${orgnr}`);
