+++ src/services/modules/types/index.ts
@@ -1,6 +1,7 @@
 import { RegisterAdresse, UstrukturertAdresse, MidlertidigAdresse, StrukturertAdresse } from "./adresse";
 import Periode from "./periode";
 import Person, { Familiemedlem } from "./person";
+import Organisasjon from "./organisasjon";
 
 export type {
   RegisterAdresse,
@@ -10,4 +11,5 @@ export type {
   Periode,
   Person,
   Familiemedlem,
+  Organisasjon,
 };
