+++ src/services/modules/types/organisasjon.ts
@@ -1,0 +1,12 @@
+import { RegisterAdresse } from "./adresse";
+
+interface Organisasjon {
+  orgnr: string;
+  navn: string;
+  forretningsadresse: RegisterAdresse;
+  postadresse: RegisterAdresse;
+  organisasjonsform: string | null;
+  oppstartdato: string | null;
+}
+
+export default Organisasjon;
